pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sklearn-diagnose"
version = "0.1.1"
description = "An intelligent diagnosis layer for scikit-learn: evidence-based model failure detection with LLM-powered summaries"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "sklearn-diagnose contributors"}
]
keywords = [
    "machine-learning",
    "scikit-learn",
    "sklearn",
    "diagnostics",
    "model-debugging",
    "llm",
    "explainability",
    "cross-validation",
    "overfitting",
    "underfitting"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "scikit-learn>=1.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    # LangChain dependencies
    "langchain>=1.2.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    # Chatbot dependencies
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/leockl/sklearn-diagnose"
Documentation = "https://github.com/leockl/sklearn-diagnose#readme"
Repository = "https://github.com/leockl/sklearn-diagnose"
Issues = "https://github.com/leockl/sklearn-diagnose/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["sklearn_diagnose*"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
